requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
mysql-connector-python
firebase-admin
gunicorn
tensorflow
pillow
python-dotenv